Add files via upload
The special feature of this site is that it also helps students to choose their career path after graduation. This is done through a questionnaire and then using the responses given by the user to decide which exam is most suitable for the user . 
This project uses both frontend as well as backend technologies.